rename: docs 관련 어노테이션 위치 이동 (#206)
